further cleaning modules and move functions into functional
