+++ corebehrt/main/finetune_cv.py
@@ -4,26 +4,23 @@ from os.path import join, split
 
 import torch
 
-from corebehrt.modules.setup.config import load_config
-from corebehrt.modules.setup.manager import ModelManager
-from corebehrt.modules.setup.loader import load_and_select_splits
-from corebehrt.modules.setup.directory import DirectoryPreparer
-from corebehrt.functional.setup.args import get_args
-from corebehrt.modules.monitoring.metric_aggregation import (
-    compute_and_save_scores_mean_std,
-)
-
-from corebehrt.functional.load import load_pids
-
 from corebehrt.functional.features.split import (
     get_n_splits_cv_pids,
     split_into_test_and_train_val_pids,
 )
+from corebehrt.functional.load import load_pids
+from corebehrt.functional.setup.args import get_args
+from corebehrt.functional.setup.data import load_and_select_splits
 from corebehrt.functional.trainer.setup import replace_steps_with_epochs
+from corebehrt.modules.monitoring.metric_aggregation import (
+    compute_and_save_scores_mean_std,
+)
 from corebehrt.modules.preparation.dataset import BinaryOutcomeDataset, PatientDataset
 from corebehrt.modules.preparation.prepare_data import DatasetPreparer
+from corebehrt.modules.setup.config import load_config
+from corebehrt.modules.setup.directory import DirectoryPreparer
+from corebehrt.modules.setup.manager import ModelManager
 from corebehrt.modules.trainer.trainer import EHRTrainer
-
 
 CONFIG_PATH = "./corebehrt/configs/finetune.yaml"
